refactor User, Patient > persons.py
